// packages/typespec-test/test/Fleet.Management/generated/typespec-ts/src/models/models.ts
export interface ArmResource extends ArmResourceBase {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface ProxyResourceBase extends ArmResource {}
// ...
export type ProvisioningState = string;
// ...
export type RegionalDataBoundary = string;
// ...
export interface DeviceGroup extends ProxyResourceBase {
  /** The resource-specific properties for this resource. */
  properties?: DeviceGroupProperties;
}

/** The properties of deviceGroup */
export interface DeviceGroupProperties {
  /** Description of the device group. */
  description?: string;
  /** Operating system feed type of the device group. */
  osFeedType?: OSFeedType;
  /** Update policy of the device group. */
  updatePolicy?: UpdatePolicy;
  /** Flag to define if the user allows for crash dump collection. */
  allowCrashDumpsCollection?: AllowCrashDumpCollection;
  /** Regional data boundary for the device group. */
  regionalDataBoundary?: RegionalDataBoundary;
  /** Deployment status for the device group. */
  readonly hasDeployment?: boolean;
  /** The status of the last operation. */
  readonly provisioningState?: ProvisioningState;
}

/** OS feed type values. */
/** "Retail", "RetailEval" */
export type OSFeedType = string;
/** Update policy values. */
/** "UpdateAll", "No3rdPartyAppUpdates" */
export type UpdatePolicy = string;
/** Allow crash dumps values. */
/** "Enabled", "Disabled" */
export type AllowCrashDumpCollection = string;

/** The type used for update operations of the DeviceGroup. */
export interface DeviceGroupUpdate {
  /** The updatable properties of the DeviceGroup. */
  properties?: DeviceGroupUpdateProperties;
}

/** The updatable properties of the DeviceGroup. */
export interface DeviceGroupUpdateProperties {
  /** Description of the device group. */
  description?: string;
  /** Operating system feed type of the device group. */
  osFeedType?: OSFeedType;
  /** Update policy of the device group. */
  updatePolicy?: UpdatePolicy;
  /** Flag to define if the user allows for crash dump collection. */
  allowCrashDumpsCollection?: AllowCrashDumpCollection;
  /** Regional data boundary for the device group. */
  regionalDataBoundary?: RegionalDataBoundary;
}

/** The response of a DeviceGroup list operation. */
export interface DeviceGroupListResult {
  /** The DeviceGroup items on this page */
  value: DeviceGroup[];
  /** The link to the next page of items */
  nextLink?: string;
}

/** Request to the action call to bulk claim devices. */
export interface ClaimDevicesRequest {
  /** Device identifiers of the devices to be claimed. */
  deviceIdentifiers: string[];
}

/** Response of the count for elements. */
export interface CountElementsResponse {
  /** Number of children resources in parent resource. */
  value: number;
}

/** Response to the action call for count devices in a catalog. */
export interface CountDevicesResponse extends CountElementsResponse {}